requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-filter==2.2.0
djangorestframework==3.10.3
freeze==1.0.10
Markdown==3.1.1
Pillow==6.1.0
psycopg2==2.8.2
pytz==2019.2
reportlab==3.5.26
six==1.12.0
sqlparse==0.3.0
